fix for jest pre 27
