+++ package.json
@@ -5,6 +5,7 @@
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
   "files": [
+    "jest-preset.js",
     "metamask",
     "dist"
   ],
@@ -36,11 +37,11 @@
     "node-stream-zip": "^1.13.0"
   },
   "devDependencies": {
+    "@jest/types": "^27.1.1",
     "@types/mocha": "^7.0.2",
     "@types/puppeteer": "^5.4.3",
     "@typescript-eslint/eslint-plugin": "^4.15.0",
     "@typescript-eslint/parser": "^4.15.0",
-    "@jest/types": "^27.1.1",
     "eslint": "^7.20.0",
     "eslint-config-prettier": "^7.2.0",
     "eslint-plugin-import": "^2.22.1",
